Move dialer to conpool package.
