+++ cmd/traffic/cmd/manager/internal/state/state.go
@@ -16,7 +16,6 @@ import (
 
 	"github.com/datawire/dlib/dlog"
 	rpc "github.com/telepresenceio/telepresence/rpc/v2/manager"
-	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/internal/conntunnel"
 	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/internal/watchable"
 	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/managerutil"
 	"github.com/telepresenceio/telepresence/v2/pkg/connpool"
@@ -542,7 +541,7 @@ func (s *State) handleTunnelMessage(ctx context.Context, pool *connpool.Pool, se
 	h, err := pool.Get(ctx, id, func(ctx context.Context, release func()) (connpool.Handler, error) {
 		switch id.Protocol() {
 		case unix.IPPROTO_TCP, unix.IPPROTO_UDP:
-			return conntunnel.NewDialer(id, server, release), nil
+			return connpool.NewDialer(id, server, release), nil
 		default:
 			return nil, fmt.Errorf("unhadled L4 protocol: %d", id.Protocol())
 		}
